code cleanup avatar-view separation to be reusable and more modular refactored backed to implement interface
