+++ app/Repositories/Interface/DeleteableEntityInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Repositories\Interface;
+
+interface DeleteableEntityInterface {
+    public function delete($params);
+}
