+++ app/Repositories/Interface/ListableEntityInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Repositories\Interface;
+
+interface ListableEntityInterface {
+    public function list($params);
+    public function count($id);
+}
