+++ resources/js/src/pages/Register.jsx
@@ -14,7 +14,7 @@ import { Link as ReactLink, useNavigate } from 'react-router-dom';
 import { useFormik } from 'formik';
 import { getRegisterFormik } from '../utils/formik-config';
 import nProgress from 'nprogress';
-import Auth from '../axios/auth';
+import Api from '../axios/account';
 
 const Register = () => {
 
@@ -31,7 +31,7 @@ const Register = () => {
     const onSubmit = async (values, helpers) => {
         nProgress.start();
         setAlert(alertdef);
-        Auth.register(values).then(res => {
+        Api.register(values).then(res => {
             const data = res.data;
             if (data.success === false) {
                 nProgress.done();
